+++ pkg/scalers/kafka_scaler_test.go
@@ -36,14 +36,6 @@ type kafkaMetricIdentifier struct {
 	name             string
 }
 
-// A complete valid metadata example for reference
-var validKafkaMetadata = map[string]string{
-	"bootstrapServers":   "broker1:9092,broker2:9092",
-	"consumerGroup":      "my-group",
-	"topic":              "my-topic",
-	"allowIdleConsumers": "false",
-}
-
 // A complete valid authParams example for sasl, with username and passwd
 var validWithAuthParams = map[string]string{
 	"sasl":     "plaintext",